Addon that adds audio reading to Japanese notes
This addon will first look for the audio reading of the selected
Japanese note using the kanji and kana. If the audio is found, it will
be downloaded to the media directory with the note's id as the filename.
The filename is then added to the audio field of the note.